Semi-working submission to linked list
Added Movie and Game subclasses for the Media. Got the Game class adding to the linked list and printing the to string.

Next Steps are to get the Movie class working and then get a function to edit the linked list item to add a rented date and then add that object to the priority queue.